refactor (Api.js): relocate to package root
* update tests and paths
* relocate test spec to match pathing
* added Api.js template mapper to package Api.js
