+++ package.json
@@ -2,7 +2,7 @@
   "name": "cordova-electron",
   "version": "2.0.0-dev",
   "description": "electron apps as a target for cordova developers",
-  "main": "bin/templates/cordova/Api.js",
+  "main": "lib/Api.js",
   "repository": "github:apache/cordova-electron",
   "bugs": "https://github.com/apache/cordova-electron/issues",
   "kewords": [
